build: Remove the usage of a local maven repository for ORT dependencies
The intention for using a local maven repository for ORT was to allow
changing the ORT version during development. [1] broke the current
approach to do this.

The better approach is to setup a local gradle configuration in
`~/.gradle/init.gradle.kts` with this content:
```kt
settingsEvaluated {
    settings.pluginManagement {
        repositories {
            mavenLocal()
            mavenCentral()
        }
    }
}
```

As the configuration in `ort-server-base-conventions.gradle.kts` is no
longer relevant for the used ORT version, remove the `mavenLocal()`
configuration.

Signed-off-by: Marcel Bochtler <marcel.bochtler@bosch.com>